--- webapp/src/components/footer.module.scss
@@ -1,15 +1,0 @@
-@use "../styles/variables/sass.scss" as *;
-
-.block {
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-  height: 2rem;
-  background-color: var(--colour1-secondary);
-}
-
-.link {
-  font-size: 1.25rem;
-  line-height: 2rem;
-  font-weight: bold;
-}
