--- webapp/src/components/header.module.scss
@@ -1,14 +1,0 @@
-@use "../styles/variables/sass.scss" as *;
-
-.block {
-  display: flex;
-  justify-content: center;
-  height: 4rem;
-  background-color: var(--colour1-secondary);
-}
-
-.logo {
-  font-size: 1.5rem;
-  line-height: 2rem;
-  font-weight: bold;
-}
