+++ webapp/src/components/layout/_index.ts
@@ -1,0 +1,2 @@
+export { Layout } from "./layout";
+export type { ILayoutProps } from "./layout";
