+++ webapp/src/components/layout/layout.module.scss
@@ -1,0 +1,47 @@
+@use "../../styles/variables/sass.scss" as *;
+
+.header {
+  display: flex;
+  justify-content: center;
+  height: 4rem;
+  background-color: var(--colour1-secondary);
+}
+
+.nav {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 1rem;
+  max-width: 1024px;
+  width: 100vw;
+  margin: 0 1rem;
+}
+
+.list {
+  display: flex;
+  justify-content: flex-start;
+  flex-grow: 0;
+}
+
+.logo {
+  font-size: 1.5rem;
+  line-height: 2rem;
+  font-weight: bold;
+}
+
+.main {
+}
+
+.footer {
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+  height: 2rem;
+  background-color: var(--colour1-secondary);
+}
+
+.link {
+  font-size: 1.25rem;
+  line-height: 2rem;
+  font-weight: bold;
+}
